+++ fxn/api/environment.py
@@ -7,6 +7,8 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import List, Optional
 
+from .api import query
+
 @dataclass(frozen=True)
 class EnvironmentVariable:
     """
@@ -18,17 +20,46 @@ class EnvironmentVariable:
     """
     name: str
     value: Optional[str]=None
+    FIELDS = f"""
+    name
+    """
+    DEFAULT_VALUE = "xxxxxxxx"
 
     @classmethod
-    def list ( # INCOMPLETE
+    def list (
         cls,
         organization: str=None,
         access_key: str=None
     ) -> List[EnvironmentVariable]:
-        pass
+        # Query
+        response = query(f"""
+            query ($input: UserInput) {{
+                user (input: $input) {{
+                    ... on User {{
+                        environmentVariables {{
+                            {cls.FIELDS}
+                        }}
+                    }}
+                    ... on Organization {{
+                        environmentVariables {{
+                            {cls.FIELDS}
+                        }}
+                    }}
+                }}
+            }}
+            """,
+            { "input": { "username": organization } if organization is not None else None },
+            access_key=access_key
+        )
+        # Create envs
+        assert response["user"] is not None, "Failed to list environment variables because user could not be found. Check that you are authenticated."
+        environments = response["user"]["environmentVariables"]
+        environments = [EnvironmentVariable(**env, value=cls.DEFAULT_VALUE) for env in environments]
+        # Return
+        return environments
 
     @classmethod
-    def create ( # INCOMPLETE
+    def create (
         cls,
         name: str,
         value: str,
@@ -49,10 +80,25 @@ class EnvironmentVariable:
         Returns:
             EnvironmentVariable: Created environment variable.
         """
-        pass
+        # Query
+        response = query(f"""
+            mutation ($input: CreateEnvironmentVariableInput!) {{
+                environment: createEnvironmentVariable (input: $input) {{
+                    {cls.FIELDS}
+                }}
+            }}
+            """,
+            { "input": { "name": name, "value": value, "organization": organization } },
+            access_key=access_key
+        )
+        # Create env
+        environment = response["environment"]
+        environment = EnvironmentVariable(**environment, value=cls.DEFAULT_VALUE)
+        # Return
+        return environment
 
     @classmethod
-    def delete ( # INCOMPLETE
+    def delete (
         cls,
         name: str,
         organization: str=None,
@@ -69,4 +115,15 @@ class EnvironmentVariable:
         Returns:
             bool: Whether the environment variable was successfully deleted.
         """
-        pass
+        # Query
+        response = query(f"""
+            mutation ($input: DeleteEnvironmentVariableInput!) {{
+                result: deleteEnvironmentVariable (input: $input)
+            }}
+            """,
+            { "input": { "name": name, "organization": organization } },
+            access_key=access_key
+        )
+        # Return
+        result = response["result"]
+        return result