+++ fxn/cli/misc.py
@@ -10,14 +10,10 @@ from webbrowser import open as open_browser
 
 from ..version import __version__
 
-def create_learn_callback (url: str) -> Callable[[bool], None]:
-    # Define callback
-    def callback (value: bool):
-        if value:
-            open_browser(url)
-            raise Exit()
-    # Return
-    return callback
+def _learn_callback (value: bool):
+    if value:
+        open_browser("https://docs.fxn.ai")
+        raise Exit()
 
 def _version_callback (value: bool):
     if value:
@@ -25,7 +21,7 @@ def _version_callback (value: bool):
         raise Exit()
 
 def cli_options (
-    learn: bool = Option(None, "--learn", callback=create_learn_callback("https://docs.fxn.ai"), help="Learn about Function."),
+    learn: bool = Option(None, "--learn", callback=_learn_callback, help="Learn about Function."),
     version: bool = Option(None, "--version", callback=_version_callback, help="Get the Function CLI version.")
 ):
     pass