+++ fxn/magic.py
@@ -13,7 +13,8 @@ class FunctionMagics (Magics):
     @line_magic
     def fxn (self, line: str):
         COMMANDS = {
-            "python": self.__python
+            "python": self.__python,
+            "image": self.__image,
         }
         args = line.split(" ")
         command = COMMANDS.get(args[0], None)
@@ -25,6 +26,10 @@ class FunctionMagics (Magics):
     def __python (self, args: List[str]):
         version = args[0]
         print(f"Predictor will use Python {version} when running on Function")
+
+    def __image (self, args: List[str]):
+        image = args[0]
+        print(f"Predictor will use base image {image} when running on Function")
 
 def load_ipython_extension (ipython: InteractiveShell):
     ipython.register_magics(FunctionMagics)